requirements.txt
numpy
pandas
matplotlib
sklearn
jupyter
parameterized
tensorflow-gpu==1.8
# for tensorboard 1.8
bleach==1.5.0